// src/dataTypes/types.ts
export interface Data {
  user: User;
  gallery: Gallery;
  art_categories: ArtCategory[];
  art_descriptions: ArtCategory[];
  art_framed: ArtCategory[];
  art_orientations: ArtCategory[];
  art_signed: ArtCategory[];
  art_sizes: ArtCategory[];
  art_themes: ArtCategory[];
  art_colors: ArtColor[];
  regions: ArtCategory[];
  countries: Country[];
}

export interface ArtCategory {
  id: number;
  name: string;
}

export interface ArtColor {
  id: number;
  name: string;
  image: Image;
}

export interface Image {
  thumbnail: string;
  small: string;
  medium: string;
  large: string;
  original: string;
}

export interface Country {
  id: number;
  name: string;
  region_id: number | null;
}

export interface Gallery {
  id: number;
  title: string;
  city: string;
  country: string;
  link: string;
  description: string;
  email: string;
  image: Image;
  artists: number;
  artworks: Artworks;
  shows: number;
  analytics: Analytics;
  dashboard_analytics: DashboardAnalytics;
}

export interface Analytics {
  views: Views;
}

export interface Views {
  all_time: AllTime;
  monthly: Monthly[];
}

export interface AllTime {
  total_views: number;
  logged_in_users: number;
  anonymous_users: number;
  by_country: ByCountry[];
}

export interface ByCountry {
  name: string;
  value: number;
}

export interface Monthly {
  month: string;
  year: string;
  data: AllTime;
}

export interface Artworks {
  total: number;
  live: number;
  for_sale: number;
  pending: number;
}

export interface DashboardAnalytics {
  artists: AnalyticsData[];
  artworks: AnalyticsData[];
  countries: AnalyticsData[];
}

export interface AnalyticsData {
  id: number;
  image?: string;
  name: string;
  view_count: number;
  percentage: number;
}

export interface User {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  mobile: string;
  facebook_id: string;
  verified: boolean;
  roleable_type: string;
  roleable_id: number;
  role: string;
  city: string | null;
  country: string | null;
  inbox_name: string;
  profession: string | null;
  buy_art: null;
  preference_range: null;
  newsletter: boolean;
  auth_token: string;
  id_verification_status: string;
  info_verified: boolean;
  interested_in: any[];
}

// Generated by https://quicktype.io

export interface Artists {
  success: boolean;
  data: Data;
  code: string;
  error: null;
}

export interface Data {
  items: ArtistItem[];
  current_count: number;
  total: number;
  limit: number;
  offset: number;
}

export interface ArtistItem {
  id: number;
  first_name: string;
  last_name: string;
  is_featured: boolean | null;
  region_id: number;
  country_id: number;
  image: Image;
  cv_file: null;
  full_name: string;
  artworks: Artworks;
  analytics: Analytics;
  likes_analytics: LikesAnalytics;
}
// ...
export interface LikesAnalytics {
  total_count: number;
}

// Generated by https://quicktype.io

export interface ArtworksRecord {
  id: number;
  title: string;
  slug: string;
  price: string;
  is_in_stock: boolean;
  price_text: string;
  year_range: null | string;
  active: boolean;
  quantity: number;
  buyer_premium: boolean;
  art_description: number;
  image: Image;
  artist: Artist;
  analytics: Analytics;
}
// ...
export interface Artist {
  id: number;
  first_name: string;
  last_name: string;
  is_featured: boolean;
  region_id: number;
  country_id: number;
  image: Image;
  cv_file: null;
  full_name: string;
}

export interface Image {
  thumbnail: string;
  small: string;
  medium: string;
  large: string;
  original: string;
}

// Generated by https://quicktype.io

export interface Conversations {
  items: ConversationItem[];
  current_count: number;
  total: number;
  limit: number;
  offset: number;
}

export interface ConversationItem {
  id: number;
  created_at: string;
  updated_at: string;
  title: string;
  html_title: string;
  first_participant: FirstParticipant;
  second_participant: SecondParticipant;
  artwork: Artwork | null;
  last_message: LastMessage;
}

export interface Artwork {
  id: number;
  title: string;
  slug: string;
  price: string;
  is_in_stock: boolean;
  price_text: string;
  year_range: string;
  active: boolean;
  quantity: number;
  buyer_premium: boolean;
  description: string;
  image: Image;
  artist: Artist;
  size: string;
  art_categories: string[];
}

export interface Artist {
  id: number;
  slug: string;
  first_name: string;
  last_name: string;
  is_featured: boolean;
  full_name: string;
  image: Image;
  cv_file: null;
}

export interface Image {
  thumbnail: string;
  small: string;
  medium: string;
  large: string;
  original: string;
}

export interface FirstParticipant {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  mobile: string;
  facebook_id: null;
  verified: boolean;
  roleable_type: string;
  roleable_id: null;
  role: string;
  city: string;
  country: string;
  inbox_name: string;
}

export interface LastMessage {
  id: number;
  participant_type: string;
  participant_id: number;
  viewed_at: null;
  created_at: string;
  updated_at: string;
  message: string;
  participant: Participant;
  attachments: any[];
}

export interface Participant {
  id: number;
  title?: string;
  slug?: string;
  city: string;
  country: string;
  link?: string;
  description?: string;
  email: string;
  inbox_name: string;
  image?: Image;
  first_name?: string;
  last_name?: string;
  mobile?: string;
  facebook_id?: null;
  verified?: boolean;
  roleable_type?: null | string;
  roleable_id?: null;
  role?: string;
}

export interface SecondParticipant {
  id: number;
  title: string;
  slug: string;
  city: string;
  country: string;
  link: string;
  description: string;
  email: string;
  inbox_name: string;
  image: Image;
}

// Generated by https://quicktype.io

export interface GallerySystemArtists {
  success: boolean;
  data: GallerySystemData;
  code: string;
  error: null;
}

export interface GallerySystemData {
  items: GallerySystemArtist[];
  current_count: number;
  total: number;
  limit: number;
  offset: number;
}

export interface GallerySystemArtist {
  id: number;
  first_name: string;
  last_name: null | string;
  is_featured: boolean | null;
  region_id: number | null;
  country_id: number | null;
  image: Image | null;
  cv_file: CvFile | null;
  full_name: string;
}

export interface CvFile {
  original: string;
}

export interface Image {
  thumbnail: string;
  small: string;
  medium: string;
  large: string;
  original: string;
}

// Generated by https://quicktype.io

export interface SingleArtist {
  success: boolean;
  data: SingleArtistData;
  code: string;
  error: null;
}

export interface SingleArtistData {
  id: number;
  first_name: string;
  last_name: string;
  is_featured: boolean;
  region_id: number;
  country_id: number;
  image: Image;
  cv_file: null;
  full_name: string;
  artworks: Artworks;
  analytics: Analytics;
  likes_analytics: LikesAnalytics;
  about: string;
  website: string;
  articles: Article[];
  key_collections: KeyCollection[];
  selected_exhibitions: SelectedExhibition[];
}

export type Article = {
  title: string;
  url: string;
};

export interface KeyCollection {
  date: string;
  museum_name: string;
  city: string;
  country_id: number;
}
// ...
export interface SelectedExhibition {
  id?: number;
  title: string;
  date: string;
  exhibition_type: string;
  location: string;
  city: string;
  country_id: number;
}

// Generated by https://quicktype.io

export interface SingleArtwork {
  success: boolean;
  data: SingleArtworkData;
  code: string;
  error: null;
}

export interface SingleArtworkData {
  id: number;
  title: string;
  slug: string;
  price: string;
  is_in_stock: boolean;
  price_text: string;
  year_range: string;
  active: boolean;
  quantity: string | number;
  buyer_premium: boolean;
  art_description: number;
  image: Image;
  artist: Artist;
  analytics: Analytics;
  height: string;
  width: string;
  length: string | null;
  depth: string | null;
  weight: string | null;
  details: string;
  shipping: string;
  edition_of: string | null;
  provenance: string | null;
  condition_report: string;
  location: string;
  is_for_charity: boolean;
  attachments: Attachment[];
  edition: string | null;
  artwork_details: string;
  size: string;
  signed: number;
  framed: number;
  region: number;
  country: number;
  art_size: number;
  art_orientation: number;
  art_categories: number[];
  art_colors: number[];
  art_themes: number[];
}
// ...
export interface Attachment {
  id: number;
  file_file_name: string;
  file_content_type: string;
  sort_order: number;
  thumbnail: string;
  small: string;
  medium: string;
  large: string;
  original: string;
}
// ...
export interface SortOrder {
  id: number;
  order: number;
}